/ ‘General/AbidArduinoTimer/AbidArduinoTimer.ino’

+++ General/AbidArduinoTimer/AbidArduinoTimer.ino
@@ -11,9 +11,9 @@ long nowtime;
 //int interval = 6*60*60;
 
 // Set the interval you want
-int hours = 10; 
-int minutes = 0;
-int seconds = 10;
+long hours = 10; 
+long minutes = 0;
+long seconds = 10;
 long interval = hours*60*60 + minutes*60 + seconds;
